+++ 4-full-mvc-framework/views/overview/index.php
@@ -22,7 +22,7 @@
                 echo '<tr class="active">';            
             }        
             echo '<td>'.$user->user_id.'</td>';            
-            echo '<td class="avatar"><img src="'.$user->user_gravatar_link.'" /></td>';
+            echo '<td class="avatar"><img src="'.$user->user_avatar_link.'" /></td>';
             echo '<td>'.$user->user_name.'</td>';
             echo '<td>'.$user->user_email.'</td>';
             echo '<td>Active: '.$user->user_active.'</td>';